+++ src/api/auth.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import { setupInterceptors, startTokenRefresh, stopTokenRefresh, refreshToken } from './commonApi';
 
 const api = axios.create({
   baseURL: 'http://localhost:8080/api',
@@ -8,107 +9,8 @@ const api = axios.create({
   withCredentials: true,
 });
 
-let refreshTimer = null;
-
-function getTokenExpiration(token) {
-  try {
-    const payload = token.split('.')[1];
-    const decoded = JSON.parse(
-      atob(payload.replace(/-/g, '+').replace(/_/g, '/'))
-    );
-    return decoded.exp ? decoded.exp * 1000 : null;
-  } catch (e) {
-    return null;
-  }
-}
-
-const refreshAccessToken = async () => {
-  try {
-    console.log('토큰 갱신 시도 중');
-    const response = await api.post('/auth/refresh');
-    const newToken = response.data.token;
-
-    localStorage.setItem('token', newToken);
-    console.log('토큰 갱신 성공:', new Date().toLocaleTimeString());
-
-    scheduleTokenRefresh(newToken);
-
-    return newToken;
-  } catch (error) {
-    console.error('토큰 갱신 실패:', error);
-    localStorage.removeItem('token');
-    window.location.href = '/auth/login';
-    throw error;
-  }
-};
-
-const scheduleTokenRefresh = (token) => {
-  if (refreshTimer) {
-    clearTimeout(refreshTimer);
-  }
-  if (!token) {
-    token = localStorage.getItem('token');
-  }
-  if (!token) return;
-
-  const exp = getTokenExpiration(token);
-  if (!exp) return;
-  const now = Date.now();
-  let msUntilRefresh = exp - now;
-  msUntilRefresh = Math.max(Math.floor(msUntilRefresh * 0.9), 5000);
-  if (msUntilRefresh < 1000) msUntilRefresh = 1000;
-
-  console.log(
-    '토큰 만료까지 남은 시간(ms):',
-    exp - now,
-    ',',
-    Math.floor((exp - now) / 1000),
-    '초'
-  );
-  console.log(Math.floor(msUntilRefresh / 1000), '초 후 토큰 갱신 예정');
-
-  refreshTimer = setTimeout(() => {
-    refreshAccessToken();
-  }, msUntilRefresh);
-};
-
-api.interceptors.response.use(
-  (response) => response,
-  async (error) => {
-    const originalRequest = error.config;
-
-    if (error.response?.status === 401 && !originalRequest._retry) {
-      originalRequest._retry = true;
-
-      try {
-        console.log('401 에러로 인한 토큰 갱신 시도');
-        const newToken = await refreshAccessToken();
-
-        originalRequest.headers['Authorization'] = `Bearer ${newToken}`;
-
-        return api(originalRequest);
-      } catch (refreshError) {
-        console.error('토큰 갱신 실패로 로그인 페이지로 이동');
-        return Promise.reject(refreshError);
-      }
-    }
-
-    return Promise.reject(error);
-  }
-);
-
-api.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem('token');
-    if (token) {
-      config.headers['Authorization'] = `Bearer ${token}`;
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  }
-);
+// 인터셉터 설정
+setupInterceptors(api);
 
 export default {
   async kakaoLogin(code) {
@@ -117,7 +19,7 @@ export default {
 
     if (data.token) {
       localStorage.setItem('token', data.token);
-      scheduleTokenRefresh(data.token);
+      startTokenRefresh(api);
       console.log('카카오 로그인 성공, 토큰 갱신 스케줄링 시작');
     }
 
@@ -130,7 +32,7 @@ export default {
 
     if (data.token) {
       localStorage.setItem('token', data.token);
-      scheduleTokenRefresh(data.token);
+      startTokenRefresh(api);
       console.log('구글 로그인 성공, 토큰 갱신 스케줄링 시작');
     }
 
@@ -178,44 +80,29 @@ export default {
     try {
       const { data } = await api.post('/auth/logout');
 
-      if (refreshTimer) {
-        clearTimeout(refreshTimer);
-        refreshTimer = null;
-      }
-
+      stopTokenRefresh();
       localStorage.removeItem('token');
       console.log('로그아웃 완료, 토큰 갱신 타이머 정리');
 
       return data;
     } catch (error) {
       console.error('로그아웃 요청 실패:', error);
-      if (refreshTimer) {
-        clearTimeout(refreshTimer);
-        refreshTimer = null;
-      }
+      stopTokenRefresh();
       localStorage.removeItem('token');
       throw error;
     }
   },
 
   async refreshToken() {
-    return await refreshAccessToken();
+    return await refreshToken(api);
   },
 
   startTokenRefresh() {
-    const token = localStorage.getItem('token');
-    if (token) {
-      scheduleTokenRefresh(token);
-      console.log('토큰 갱신 스케줄링 시작');
-    }
+    startTokenRefresh(api);
   },
 
   stopTokenRefresh() {
-    if (refreshTimer) {
-      clearTimeout(refreshTimer);
-      refreshTimer = null;
-      console.log('토큰 갱신 스케줄링 중지');
-    }
+    stopTokenRefresh();
   },
 };
